Fixed dupes. Updated texture. Chesty more reliable.
Due to the Sceptres mechanics there were a lot of dupes that could be
done if you simply tossed the sceptre away and continued to try to use
Chesty, it would still allow you to take the items and because you no
longer had the sceptre it would not be able to save. Now it prevents you
from withdrawling/depositing items if you do not have the valid sceptre
for that Chesty.

Really hoping there aren't any other dupes.

Additionally Chesty would sometimes despawn while moving the Sceptre
around in your inventory. I found out how to tell if the Sceptre is
being held while accessing the inventory and should no longer have
problems with that.