IronChest mod support!
Also fixed a likely problem with many people using the mod at once in
regards of "lastInteractChesty", had to make my own packet for opening
the interface since the preexisting implementation is meant for tile
entities/blocks and not for moving entities that likely won't be there
when the packet was actually received.

There's lots of testing still to be done, hopefully no problems, though.